ADD undo remove todo

+++ src/mixins/removeTodo.js
@@ -1,26 +1,43 @@
 import {
   db
 } from '@/firebase'
-import {
-  toast
-} from '@/mixins/toast'
 
 export const removeTodo = {
   methods: {
-    removeTodo(todo) {
+    data: _ => ({
+      isRemoved: false,
+    }),
+
+    removeTodo(todo, delay) {
+      todo.removed = true
+      this.isRemoved = true
+
+      this.$toasted.show(`${todo.text} removed`, {
+        theme: 'primary',
+        position: 'bottom-center',
+        duration: delay,
+        action: {
+          text: 'UNDO',
+          onClick: (e, toastObject) => {
+            todo.removed = false
+            toastObject.goAway(0)
+            return this.isRemoved = false
+          }
+        },
+      })
+
       setTimeout(() => {
-        db.collection(todo.uid).doc(todo.id).delete()
-          .then(() => {
-            this.toast('Todo removed', 2000, 'Close')
-          })
-          .catch(error => {
-            console.log(error)
-          })
-      }, 0)
+        if (this.isRemoved) {
+          db.collection(todo.uid).doc(todo.id).delete()
+            .then((response) => {
+              console.log(response)
+            })
+            .catch(error => {
+              console.log(error)
+            })
+        }
+      }, delay)
     },
   },
 
-  mixins: [
-    toast,
-  ],
 }
